requirements.txt
# Django and related
Django==5.1.5
djangorestframework==3.16.0
django-storages==1.14.6

# TensorFlow and related
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.15.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tf_keras==2.15.1

# Machine Learning and Data Science
scikit-learn==1.6.1
scipy==1.15.3
numpy==1.26.4
pandas==2.2.3
joblib==1.5.0
h5py==3.13.0
scikit-learn==1.6.1
ml-dtypes==0.2.0

# Google and AWS SDKs
boto3==1.38.14
botocore==1.38.14
s3transfer==0.12.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.71.0

# Miscellaneous utilities
python-dotenv==1.1.0
psycopg2==2.9.10
requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2
idna==3.10
jmespath==1.0.1
urllib3==2.4.0

# Authentication and Security
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
six==1.17.0

# Formatting and utilities
Markdown==3.8
MarkupSafe==3.0.2
Werkzeug==3.1.3
termcolor==3.1.0
cachetools==5.5.2
flatbuffers==25.2.10
gast==0.6.0
opt_einsum==3.4.0
packaging==25.0
typing_extensions==4.13.2
tzdata==2024.2

# Development tools
absl-py==2.2.2
astunparse==1.6.3
libclang==18.1.1
sqlparse==0.5.3
python-dateutil==2.9.0.post0
